+++ app/filters/contact_method_filter.rb
@@ -8,6 +8,10 @@ class ContactMethodFilter < BaseFilter
 
   def filter(scope)
     return super unless parameters
+
+    # We may want to add contact method logic for community resources. For now, this effectively allows all
+    return super unless scope.respond_to? :person_id
+
     scope.joins(:person).where(people: {preferred_contact_method: parameters.keys})
   end
 end