+++ app/models/community_resource.rb
@@ -21,6 +21,8 @@ class CommunityResource < ApplicationRecord
 
   scope :approved,       -> { where(is_approved: true) }
   scope :pending_review, -> { where(is_approved: false) }
+  # TODO: add tests for this?
+  scope :in_service_areas, ->(ids) { joins(:service_areas).where(service_areas: {id: ids}).distinct }
 
   def self.published
     before_now = DateTime.new..Time.current
@@ -31,11 +33,18 @@ class CommunityResource < ApplicationRecord
     )
   end
 
+  def title; description; end
+  def self.matchable; published; end
+
   def published?
     now = Time.current
     is_approved &&
       (publish_from.present? ? publish_from <= now : true) &&
       (publish_until.nil? || now < publish_until)
+  end
+
+  def categories_for_tags
+    Category.where(name: tag_list)
   end
 
   def all_tags_unique
@@ -45,6 +54,27 @@ class CommunityResource < ApplicationRecord
   def all_tags_to_s
     all_tags_unique.join(', ')
   end
+
+  def preferred_contact_method
+    # TODO: This is a hack that makes things work for now
+    # The test creates a contact method that will match this
+    # and the dev db seeding creates a couple, too
+    ContactMethod.method_name('call').last
+  end
+
+  def type
+    "Community Resource"
+  end
+
+  def inexhaustible
+    true
+  end
+
+  def urgency_level_id
+    UrgencyLevel::TYPES.last.id
+  end
+
+  def person; end
 end
 
 # == Schema Information